pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scoreforge"
version = "1.0.0"
description = "ScoreForge - Sheet Music Scanner and Editor with OMR, MIDI, MusicXML, and PDF support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Tim Kenobi", email = "tim@scoreforge.app"}
]
keywords = ["music", "omr", "midi", "musicxml", "sheet-music", "optical-music-recognition", "scoreforge", "notation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: MIDI",
    "Topic :: Multimedia :: Graphics :: Capture :: Scanners",
]

dependencies = [
    "music21>=9.1.0",
    "mido>=1.3.0",
    "PySide6>=6.6.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "PyMuPDF>=1.23.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
omr = [
    "oemer>=0.1.8",
]
pdf = [
    "abjad>=3.19",
]
midi = [
    "python-rtmidi>=1.5.0",
]
playback = [
    "pygame>=2.5.0",
]
notation = [
    "verovio>=4.0.0",
]
all = [
    "oemer>=0.1.8",
    "abjad>=3.19", 
    "python-rtmidi>=1.5.0",
    "verovio>=4.0.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-qt>=4.2.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
scoreforge = "sheet_music_scanner.main:main"

[project.gui-scripts]
scoreforge-gui = "sheet_music_scanner.main:main"

[project.urls]
Homepage = "https://github.com/TimKenobi/ScoreForge"
Documentation = "https://github.com/TimKenobi/ScoreForge#readme"
Repository = "https://github.com/TimKenobi/ScoreForge"
"Buy Me a Coffee" = "https://coff.ee/timkenobi"

[tool.setuptools.packages.find]
where = ["."]
include = ["sheet_music_scanner*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
